apply rows redux and refactor table row
