+++ src/components/pages/order/bonus-formula.tsx
@@ -1,0 +1,53 @@
+import { useSelector } from 'react-redux';
+
+import { selector } from '@/redux';
+
+export const BonusFormula = () => {
+  const { order } = useSelector(selector.order);
+  const { rows } = useSelector(selector.rows);
+
+  const quantity = rows.length;
+  const bonus = (order.shipFee - order.discount) / quantity;
+  const roundedBonus = Math.ceil(bonus / 100) * 100;
+
+  return (
+    <div className="flex gap-1">
+      <div className="text-red-400">
+        <div className="font-semibold">bonus</div>
+        <div>{roundedBonus.toLocaleString('en-US')}</div>
+      </div>
+      <div>
+        <div>=</div>
+        <div>=</div>
+      </div>
+      <div>
+        <div>(</div>
+        <div>(</div>
+      </div>
+      <div className="min-w-fit">
+        <div className="font-semibold">ship fee</div>
+        <div>{Number(order.shipFee).toLocaleString('en-US')}</div>
+      </div>
+      <div>
+        <div>-</div>
+        <div>-</div>
+      </div>
+      <div>
+        <div className="font-semibold">discount</div>
+        <div>{Number(order.discount).toLocaleString('en-US')}</div>
+      </div>
+      <div>
+        <div>)</div>
+        <div>)</div>
+      </div>
+      <div>
+        <div>/</div>
+        <div>/</div>
+      </div>
+      <div>
+        <div>quantity</div>
+        <div>{quantity}</div>
+      </div>
+    </div>
+  );
+};
