+++ src/components/pages/order/delete-row-button.tsx
@@ -1,0 +1,47 @@
+import { CheckIcon, TrashIcon, XMarkIcon } from '@heroicons/react/24/outline';
+import { deleteDoc, doc } from 'firebase/firestore';
+import { useState } from 'react';
+import { useSelector } from 'react-redux';
+
+import { db } from '@/firebase';
+import { useCheckClickOutside } from '@/hooks';
+import { selector } from '@/redux';
+import type { DrinkTableRow } from '@/types';
+
+export const DeleteRowButton = ({ row }: { row: DrinkTableRow }) => {
+  const { order } = useSelector(selector.order);
+
+  const [isDropdown, setIsDropdown] = useState(false);
+
+  const deleteRowButtonRef = useCheckClickOutside(() => setIsDropdown(false));
+
+  const _deleteRow = async () => {
+    const docRef = doc(db, 'orders', order.id, 'rows', row.id);
+    await deleteDoc(docRef);
+  };
+
+  return (
+    <div ref={deleteRowButtonRef} className="relative">
+      <button
+        type="button"
+        onClick={() => setIsDropdown(true)}
+        className="rounded-md bg-gray-400 p-1 hover:bg-gray-500"
+      >
+        <TrashIcon className="h-3 w-3" />
+      </button>
+      {isDropdown ? (
+        <div className="absolute -left-7 top-6 z-10 flex gap-1 rounded-md bg-white p-1">
+          <button className="rounded-md bg-gray-200 p-1" onClick={_deleteRow}>
+            <CheckIcon className="h-3 w-3" />
+          </button>
+          <button
+            className="rounded-md bg-gray-200 p-1"
+            onClick={() => setIsDropdown(false)}
+          >
+            <XMarkIcon className="h-3 w-3" />
+          </button>
+        </div>
+      ) : null}
+    </div>
+  );
+};
