+++ src/components/pages/order/index.ts
@@ -1,9 +1,16 @@
+export * from './bonus-formula';
 export * from './calculate-total';
+export * from './delete-row-button';
+export * from './give-heart';
 export * from './menu-dropdown';
+export * from './offered-by-formula';
+export * from './options-dropdown';
 export * from './shared-link';
 export * from './shop-owner';
+export * from './show-formula';
 export * from './table';
 export * from './table-add-row-button';
 export * from './table-header';
 export * from './table-row';
+export * from './transfer-formula';
 export * from './transfer-info';
